+++ src/plugins/server.ts
@@ -120,7 +120,8 @@ export function spawnServer(
         }));
     }
 
-    app.use(collectionsPath, ExpressPouchDB(pseudo));
+    const pouchApp = ExpressPouchDB(pseudo);
+    app.use(collectionsPath, pouchApp);
 
     let server = null;
     if (startServer) {
@@ -130,6 +131,7 @@ export function spawnServer(
 
     return {
         app,
+        pouchApp,
         server
     };
 }